updating icons and dialogs
git-svn-id: file:///home/roberto/Git/svn@11823 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: edea1b8e30a81aa7936434359532af78dab09e8d

+++ trunk/xmipp/java/src/xmipp/viewer/models/MetadataTable.java
@@ -158,7 +158,7 @@ public class MetadataTable extends MetadataGallery {
 				md.setValueDouble(label, ((Double) value).doubleValue(), id);
 				break;
 			case MetaData.LABEL_SIZET:
-				md.setValueInt(label, ((Integer) value).intValue(), id);
+				md.setValueInt(label, ((Long) value).intValue(), id);
 				break;
 			case MetaData.LABEL_STRING:
 			case MetaData.LABEL_VECTOR_DOUBLE:
